+++ src/Symfony/Bundle/FrameworkBundle/Console/Descriptor/TextDescriptor.php
@@ -261,10 +261,13 @@ class TextDescriptor extends Descriptor
             $description[] = '<comment>Tags</comment>             -';
         }
 
-        $description[] = sprintf('<comment>Scope</comment>            %s', $definition->getScope());
+        $description[] = sprintf('<comment>Scope</comment>            %s', $definition->getScope(false));
         $description[] = sprintf('<comment>Public</comment>           %s', $definition->isPublic() ? 'yes' : 'no');
         $description[] = sprintf('<comment>Synthetic</comment>        %s', $definition->isSynthetic() ? 'yes' : 'no');
         $description[] = sprintf('<comment>Lazy</comment>             %s', $definition->isLazy() ? 'yes' : 'no');
+        if (method_exists($definition, 'isShared')) {
+            $description[] = sprintf('<comment>Shared</comment>           %s', $definition->isShared() ? 'yes' : 'no');
+        }
         if (method_exists($definition, 'isSynchronized')) {
             $description[] = sprintf('<comment>Synchronized</comment>     %s', $definition->isSynchronized(false) ? 'yes' : 'no');
         }
@@ -336,33 +339,16 @@ class TextDescriptor extends Descriptor
 
         $this->writeText($this->formatSection('event_dispatcher', $label)."\n", $options);
 
-        $registeredListeners = $eventDispatcher->getListeners($event);
+        $registeredListeners = $eventDispatcher->getListeners($event, true);
 
         if (null !== $event) {
             $this->writeText("\n");
-            $table = new Table($this->getOutput());
-            $table->getStyle()->setCellHeaderFormat('%s');
-            $table->setHeaders(array('Order', 'Callable'));
-
-            foreach ($registeredListeners as $order => $listener) {
-                $table->addRow(array(sprintf('#%d', $order + 1), $this->formatCallable($listener)));
-            }
-
-            $table->render();
+            $this->renderEventListenerTable($registeredListeners);
         } else {
             ksort($registeredListeners);
             foreach ($registeredListeners as $eventListened => $eventListeners) {
                 $this->writeText(sprintf("\n<info>[Event]</info> %s\n", $eventListened), $options);
-
-                $table = new Table($this->getOutput());
-                $table->getStyle()->setCellHeaderFormat('%s');
-                $table->setHeaders(array('Order', 'Callable'));
-
-                foreach ($eventListeners as $order => $eventListener) {
-                    $table->addRow(array(sprintf('#%d', $order + 1), $this->formatCallable($eventListener)));
-                }
-
-                $table->render();
+                $this->renderEventListenerTable($eventListeners);
             }
         }
     }
@@ -373,6 +359,26 @@ class TextDescriptor extends Descriptor
     protected function describeCallable($callable, array $options = array())
     {
         $this->writeText($this->formatCallable($callable), $options);
+    }
+
+    /**
+     * @param array $array
+     */
+    private function renderEventListenerTable(array $eventListeners)
+    {
+        $table = new Table($this->getOutput());
+        $table->getStyle()->setCellHeaderFormat('%s');
+        $table->setHeaders(array('Order', 'Callable', 'Priority'));
+
+        krsort($eventListeners);
+        $order = 1;
+        foreach ($eventListeners as $priority => $listeners) {
+            foreach ($listeners as $listener) {
+                $table->addRow(array(sprintf('#%d', $order++), $this->formatCallable($listener), $priority));
+            }
+        }
+
+        $table->render();
     }
 
     /**